% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kwresearch.R
\name{accentize_queries}
\alias{accentize_queries}
\title{Tries to add accents (´ˇ¨ etc.) where they should be}
\usage{
accentize_queries(x, vol)
}
\arguments{
\item{x}{Character vector of queries.}

\item{vol}{Character vector of search volumes corresponding to queries in x.}
}
\value{
Character vector of all queries in x, with accents added to some
of them.
}
\description{
Tries to add accents (´ˇ¨ etc.) where they should be
}
