% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{extract_pattern}
\alias{extract_pattern}
\title{Extracts a regex pattern from a character vector (vectorized)}
\usage{
extract_pattern(x, pattern)
}
\arguments{
\item{x}{A character vector to extract pattern from.}

\item{pattern}{A regex pattern. If it contains capture groups, they all will
be extracted.}
}
\value{
A character vector of the same length as x.
}
\description{
Extracts a regex pattern from a character vector (vectorized)
}
\keyword{internal}
