% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_mm.R
\name{import_mm_file}
\alias{import_mm_file}
\title{Imports a single MarketingMiner CSV file}
\usage{
import_mm_file(path, quiet = FALSE)
}
\arguments{
\item{path}{Full path and name of the imported file.}

\item{quiet}{If TRUE prints no messgaes.}
}
\value{
A tibble with columns: query, volume, cpc, input, source. Rows are
just copied from the CSV file without deduplication or any other
modification.
}
\description{
Imports a single MarketingMiner CSV file
}
\keyword{internal}
