% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{join_labels}
\alias{join_labels}
\title{Deduplicates and joins multiple label values}
\usage{
join_labels(x, y, sep = ",")
}
\arguments{
\item{x}{A character vector.}

\item{y}{A character vector.}

\item{sep}{A character vector of length 1.}
}
\value{
Character vector of the same length as x. Unique label values of x and y joined together.
}
\description{
Deduplicates and joins multiple label values
}
