% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{join_patterns}
\alias{join_patterns}
\title{Deduplicates regex patterns and joins them by "|"}
\usage{
join_patterns(patterns)
}
\arguments{
\item{patterns}{A character vector of regex patterns.}
}
\value{
A character vector of length 1.
}
\description{
Deduplicates regex patterns and joins them by "|"
}
\keyword{internal}
