% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex.R
\name{kwr_build_regex}
\alias{kwr_build_regex}
\title{Creates regular expression from a single string}
\usage{
kwr_build_regex(x, and = NULL)
}
\arguments{
\item{x}{Input string.}

\item{and}{Optional string that should be matched as well.}
}
\value{
Regular expression.
}
\description{
The function does the following two things:
}
\details{
\enumerate{
\item Accented letters in a string are replaced with a character class
containing both the accented letter and the same letter without accent.
\item If the \code{and} argument is provided, the resulting regex will match both
strings in any order.
}
}
\examples{
kwr_build_regex("vajíčko")

kwr_build_regex(x = "abc", and = "xyz")
}
