% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kwresearch.R
\name{kwr_classified_queries}
\alias{kwr_classified_queries}
\title{Outputs classified queries}
\usage{
kwr_classified_queries(kwr)
}
\arguments{
\item{kwr}{A kwresearch object.}
}
\value{
A tibble.
}
\description{
Outputs classified queries
}
\examples{
queries <- data.frame(
  query = c("seo", "keyword research"),
  volume = c(1000, 500)
)
kwr <- kwresearch(queries)
\dontrun{
kwr |>
  kwr_classify("C:/Dev/R/Public/keyworder/tests/test-data/recipes.yml") |>
  kwr_classified_queries()
}
}
