% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kwresearch.R
\name{kwr_clean_queries}
\alias{kwr_clean_queries}
\title{Outputs cleaned (normalized and accentized) queries}
\usage{
kwr_clean_queries(kwr, q = NULL)
}
\arguments{
\item{kwr}{A \code{\link{kwresearch}} object.}

\item{q}{An optional regular expression to filter the output queries.}
}
\value{
A tibble.
}
\description{
Outputs cleaned (normalized and accentized) queries
}
\examples{
queries <- data.frame(
  query = c("seo", "keyword research"),
  volume = c(1000, 500)
)
kwr <- kwresearch(queries)
kwr_clean_queries(kwr)
}
