% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngrams.R
\name{kwr_collocations}
\alias{kwr_collocations}
\title{Finds collocations, i.e. multiword phrases that are more likely than their
single words}
\usage{
kwr_collocations(kwr, min_volume_prop = 0.5, min_n = 2)
}
\arguments{
\item{kwr}{A kwresearch object, which clean queries will be used.}

\item{min_volume_prop}{Minimum proportion.}

\item{min_n}{Minimum number of queries. Only the n-grams with at least this
number of queries will be included.}
}
\value{
A tibble of n-grams with a basic stats (nuber of queries and sum of
search volumes). The n-grams are orderd descendingly by number of queries
and search volume. Use dplyr::arrange to change order.
}
\description{
N-grams fully contained in another n-gram (number of queries
must be the same) are filter out from the result.
}
