% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kwresearch.R
\name{kwr_dimension_names}
\alias{kwr_dimension_names}
\title{Gets the names of all dimensions}
\usage{
kwr_dimension_names(df)
}
\arguments{
\item{df}{A data frame to get dimension names from.}
}
\value{
A vector of names.
}
\description{
Gets the names of all dimensions
}
\examples{
queries <- data.frame(
  query = c("seo", "keyword research"),
  volume = c(1000, 500)
)
kwr <- kwresearch(queries)

recipe_file <- file.path(tempdir(), "my-recipes.yml")
kwr_add_pattern(
  pattern = "seo",
  recipe_file = recipe_file,
  recipe_type = "label",
  dim_name = "my_label"
)
kwr <- kwr_classify(kwr, recipe_file)
kwr |>
  kwr_queries() |>
  kwr_dimension_names()

file.remove(recipe_file)
}
