% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kwresearch.R
\name{kwr_import}
\alias{kwr_import}
\title{Imports queries to a kwresearch object}
\usage{
kwr_import(kwr, queries)
}
\arguments{
\item{kwr}{An empty \code{\link{kwresearch}} object.}

\item{queries}{A data frame with at least one column \emph{query} and
possibly with additional columns \emph{volume}, \emph{cpc}, \emph{input}
and \emph{source}.}
}
\value{
The provided kwresearch object with imported queries.
}
\description{
Imports queries to a kwresearch object
}
\examples{
queries <- data.frame(
  query = c("seo", "keyword research"),
  volume = c(1000, 500)
)
kwr <- kwresearch() |>
  kwr_import(queries)
}
