% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune.R
\name{kwr_prune_long_queries}
\alias{kwr_prune_long_queries}
\title{From pruned data removes queries longer than a specific number of characters}
\usage{
kwr_prune_long_queries(kwr, longer_than = 60)
}
\arguments{
\item{kwr}{A kwresearch object.}

\item{longer_than}{A number.}
}
\value{
A kwresearch object with pruned data and status = 'pruned'.
}
\description{
From pruned data removes queries longer than a specific number of characters
}
