% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopwords.R
\name{kwr_remove_stopwords}
\alias{kwr_remove_stopwords}
\title{Removes stopwords from a dataset, e.g. n-grams}
\usage{
kwr_remove_stopwords(df, stopwords = kwr_stopwords())
}
\arguments{
\item{df}{A data frame with words or n-grams in the 1st column.}

\item{stopwords}{A vector of words (or n-grams) to be removed. It can be the
output of the kwr_stopwords function.}
}
\value{
The same data set as df with the rows matching stopwords removed.
}
\description{
Removes stopwords from a dataset, e.g. n-grams
}
