% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kwresearch.R
\name{kwr_removed_queries}
\alias{kwr_removed_queries}
\title{Outputs queries removed by kwr_prune}
\usage{
kwr_removed_queries(kwr, q = NULL)
}
\arguments{
\item{kwr}{A \code{\link{kwresearch}} object.}

\item{q}{An optional regular expression to filter the output queries.}
}
\value{
A tibble with queries removed by \code{\link{kwr_prune}}.
}
\description{
Outputs queries removed by kwr_prune
}
\examples{
queries <- data.frame(
  query = c("seo", "keyword research", "nonsense"),
  volume = c(1000, 500, 10)
)
kwr <- kwresearch(queries)

recipe_file <- file.path(tempdir(), "my-recipes.yml")
kwr_add_pattern(
  pattern = "nonsense",
  recipe_file = recipe_file,
  recipe_type = "remove"
)
kwr <- kwr_prune(kwr, recipe_file)
kwr_removed_queries(kwr)

file.remove(recipe_file)
}
