% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kwresearch.R
\name{kwr_show_dimension}
\alias{kwr_show_dimension}
\title{Outputs frequency table of a given dimension (label or flag)}
\usage{
kwr_show_dimension(kwr, column)
}
\arguments{
\item{kwr}{A kwresearch object, whose queries are already classified.}

\item{column}{A dimension (column) as a name, not a string.}
}
\value{
A tibble.
}
\description{
Outputs frequency table of a given dimension (label or flag)
}
