% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kwresearch.R
\name{kwr_source_queries}
\alias{kwr_source_queries}
\title{Outputs raw, source queries}
\usage{
kwr_source_queries(kwr)
}
\arguments{
\item{kwr}{A kwresearch object.}
}
\value{
A tibble.
}
\description{
Outputs raw, source queries
}
\examples{
queries <- data.frame(
  query = c("seo", "keyword research"),
  volume = c(1000, 500)
)
kwr <- kwresearch(queries)
kwr |> kwr_source_queries()
}
