% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopwords.R
\name{kwr_stopwords}
\alias{kwr_stopwords}
\title{Return stopwords as a character vector}
\usage{
kwr_stopwords(lang = "cs")
}
\arguments{
\item{lang}{Language abbreviation. Only Czech (cs) is implemented now.}
}
\value{
Character vector of stopwords.
}
\description{
Return stopwords as a character vector
}
