% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex.R
\name{kwr_test_regex}
\alias{kwr_test_regex}
\title{Explore a regex pattern applied on queries}
\usage{
kwr_test_regex(kwr, pattern)
}
\arguments{
\item{kwr}{A kwresearch object.}

\item{pattern}{A single regular experession.}
}
\value{
A list of data frames.
}
\description{
Explore a regex pattern applied on queries
}
