% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{kwr_use_recipes}
\alias{kwr_use_recipes}
\title{Adds new classification recipes to a kwresearch object}
\usage{
kwr_use_recipes(kwr, recipe_file)

kwr_use_recipes(kwr, recipe_file)
}
\arguments{
\item{kwr}{A kwresearch object.}

\item{recipe_file}{A path to a recipe file in YAML format.}
}
\value{
The input kwresearch object with updated classification recipes. If
the kwresearch object already contained recipes, they are merged with the
new ones.

The input kwresearch object with updated classification recipes. If
the kwresearch object already contained recipes, they are merged with the
new ones.
}
\description{
Adds new classification recipes to a kwresearch object

Adds new classification recipes to a kwresearch object
}
