% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kwresearch.R
\name{kwr_use_stopwords}
\alias{kwr_use_stopwords}
\title{Set a stopword list to use with n-gram functions}
\usage{
kwr_use_stopwords(kwr, stopwords)
}
\arguments{
\item{kwr}{A kwresearch object.}

\item{stopwords}{Stopwords as a character vector. You can use the
kwr_stopwords() function. These words will be removed from unigram
listings.}
}
\value{
A kwresearch object with stopwords set.
}
\description{
Set a stopword list to use with n-gram functions
}
