% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kwresearch.R
\name{kwresearch}
\alias{kwresearch}
\title{Creates an object of the kwresearch class}
\usage{
kwresearch(queries = NULL, accentize = TRUE, normalize = TRUE)
}
\arguments{
\item{queries}{Queries to import as a data frame with at least two columns:
\emph{query} (string) and \emph{volume} (numeric). Optionally the data
frame can contain three additional columns: \emph{cpc} (double),
\emph{input} (char) and \emph{source} (char). If you don't provide this
argument, you can import queries later with \code{\link{kwr_import}} or
\code{\link{kwr_import_mm}}.}

\item{accentize}{The import functions tries to add correct accents (diacritic
marks) to queries without them.}

\item{normalize}{The import functions tries to unite queries, which differ
only by order of words.}
}
\value{
An object of the kwresearch class.
}
\description{
An object of the kwresearch class is a home to your keyword
research. You can import datasets into it, research queries, classify them
and export the final keyword analysis.
}
\examples{
# create an empty object

kwr <- kwresearch()

# create an object and import queries

queries <- data.frame(
  query = c("seo", "keyword research"),
  volume = c(1000, 500)
)
kwr <- kwresearch(queries)
}
