% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{process_value}
\alias{process_value}
\title{Processes a single value of a recipe of the type label}
\usage{
process_value(df, value, name, quiet = FALSE)
}
\arguments{
\item{df}{A data frame from a kwr object (either clean_data, or
classified_data).}

\item{value}{A label value (character).}

\item{name}{A label name (character).}

\item{quiet}{If TRUE prints no messages.}
}
\value{
A data frame with updated classification.
}
\description{
Processes a single value of a recipe of the type label
}
\keyword{internal}
