% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{read_recipes}
\alias{read_recipes}
\title{Reads classification recipes from a YAML file}
\usage{
read_recipes(path)
}
\arguments{
\item{path}{Path to YAML file.}
}
\value{
Classification recipes (a list).
}
\description{
Reads classification recipes from a YAML file
}
