% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{set_flag}
\alias{set_flag}
\title{Sets a classification flag (logical value)}
\usage{
set_flag(df, name, pattern, negate = FALSE)
}
\arguments{
\item{df}{A data frame from a kwr object (either clean_data, or
classified_data).}

\item{name}{A name of the flag.}

\item{pattern}{A regex pattern (character vector of length 1).}

\item{negate}{If TRUE, queries that don't match the pattern will be flagged
as TRUE.}
}
\value{
A data frame with the flag updated.
}
\description{
Sets a classification flag (logical value)
}
