% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{set_label}
\alias{set_label}
\title{Sets a classification label (character)}
\usage{
set_label(df, name, pattern, value = NULL, exclude = NULL)
}
\arguments{
\item{df}{A data frame from a kwr object (either clean_data, or
classified_data).}

\item{name}{A name of the label.}

\item{pattern}{A regex pattern (character vector of length 1). The label is
set to matched queries only.}

\item{value}{An optional value of the label (character).}

\item{exclude}{A regex pattern (character vector of length 1). The label is
never set to matched queries.}
}
\value{
A data frame with the label updated.
}
\description{
Sets a classification label (character)
}
\keyword{internal}
