% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keywordr-package.R
\docType{package}
\name{keywordr-package}
\alias{keywordr}
\alias{keywordr-package}
\title{keywordr: Research and clasify search engine queries in R!}
\description{
A framework for a keyword research in R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MarekProkop/keywordr}
  \item Report bugs at \url{https://github.com/MarekProkop/keywordr/issues}
}

}
\author{
\strong{Maintainer}: Marek Prokop \email{mprokop@prokopsw.cz}

}
\keyword{internal}
