% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kwresearch.R
\name{kwr_classified_queries}
\alias{kwr_classified_queries}
\title{Outputs classified queries}
\usage{
kwr_classified_queries(kwr, q = NULL)
}
\arguments{
\item{kwr}{A \code{\link{kwresearch}} object.}

\item{q}{An optional regular expression to filter the output queries.}
}
\value{
A tibble.
}
\description{
Outputs classified queries
}
\examples{
queries <- data.frame(
  query = c("seo", "keyword research"),
  volume = c(1000, 500)
)
kwr <- kwresearch(queries)

recipe_file <- file.path(tempdir(), "my-recipes.yml")
kwr_add_pattern(
  pattern = "seo",
  recipe_file = recipe_file,
  recipe_type = "label",
  dim_name = "my_label"
)
kwr <- kwr_classify(kwr, recipe_file)
kwr_classified_queries(kwr)

file.remove(recipe_file)
}
