% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{kwr_classify}
\alias{kwr_classify}
\title{Classifies queries based on rules in a YAML file}
\usage{
kwr_classify(kwr, recipe_file, quiet = FALSE)
}
\arguments{
\item{kwr}{A kwr object containg queries to be classified.}

\item{recipe_file}{A path to a YAML file with classification rules.}

\item{quiet}{If TRUE prints no messgaes.}
}
\value{
A kwr object where the queries are classified. Existing
classification is preserved or updated, new is added.
}
\description{
Classifies queries based on rules in a YAML file
}
\examples{
\dontrun{
kwr <- kwr |> kwr_classify("recipe.yml")
}
}
