% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kwresearch.R
\name{kwr_clean_queries}
\alias{kwr_clean_queries}
\title{Outputs cleaned (normalized and accentized) queries}
\usage{
kwr_clean_queries(kwr)
}
\arguments{
\item{kwr}{A kwresearch object.}
}
\value{
A tibble.
}
\description{
Outputs cleaned (normalized and accentized) queries
}
\examples{
queries <- data.frame(
  query = c("seo", "keyword research"),
  volume = c(1000, 500)
)
kwr <- kwresearch(queries)
kwr |> kwr_clean_queries()
}
