% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngrams.R
\name{kwr_collocations}
\alias{kwr_collocations}
\title{Finds collocations, i.e. multiword phrases that are more likely than their
single words}
\usage{
kwr_collocations(x, min_volume_prop = 0.5, min_n = 2, quanteda = TRUE)
}
\arguments{
\item{x}{A \code{\link{kwresearch}} object, which queries will be
collocations calculated from, or a data frame of queries and volume.}

\item{min_volume_prop}{Minimum proportion.}

\item{min_n}{Minimum number of queries. Only the n-grams with at least this
number of queries will be included.}

\item{quanteda}{If TRUE, will use the 'quanteda' package, which is
recommended.}
}
\value{
A tibble of n-grams with a basic stats (number of queries and sum of
search volumes). The n-grams are ordered descendingly by number of queries
and search volume. Use dplyr::arrange to change order.
}
\description{
Finds collocations, i.e. multiword phrases that are more likely than their
single words
}
