% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kwresearch.R
\name{kwr_dimension_names}
\alias{kwr_dimension_names}
\title{Gets names of all dimensions (flags an labels)}
\usage{
kwr_dimension_names(df)
}
\arguments{
\item{df}{A data frame to get dimension names from.}
}
\value{
A vector of names.
}
\description{
Gets names of all dimensions (flags an labels)
}
