% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kwresearch.R
\name{kwr_import}
\alias{kwr_import}
\title{Imports queries to a kwresearch object}
\usage{
kwr_import(kwr, queries)
}
\arguments{
\item{kwr}{An empty kwresearch object.}

\item{queries}{A data frame with at least one column query and possibly with
additional colums volume, cpc, imput and source.}
}
\value{
The provided kwresearch object with imported queries.
}
\description{
Imports queries to a kwresearch object
}
\examples{
kwr <- kwresearch()

queries <- data.frame(
  query = c("seo", "keyword research"),
  volume = c(1000, 500)
)
kwr <- kwr_import(kwr, queries)
}
