% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_mm.R
\name{kwr_import_mm}
\alias{kwr_import_mm}
\title{Imports Marketing Miner CSV file(s) to an empty kwresearch object}
\usage{
kwr_import_mm(kwr, path, quiet = FALSE)
}
\arguments{
\item{kwr}{An empty kwresearch object.}

\item{path}{Either a single character string containing a path to a folder,
or a character vector, where each item is a path to a CSV file.}

\item{quiet}{If TRUE prints no messages.}
}
\value{
The provided kwresearch object with imported queries.
}
\description{
Imports Marketing Miner CSV file(s) to an empty kwresearch object
}
\examples{
# Import 2 files:
\dontrun{
kwr <- kwresearch() |>
  kwr_import_mm(c("f1.csv", "f2.csv"))
}

# Import all CSV files in th data-raw folder
\dontrun{
kwr <- kwresearch() |>
  kwr_import_mm("data-raw")
}
}
