% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune.R
\name{kwr_long_queries}
\alias{kwr_long_queries}
\title{Outputs queries longer than a specific number of characters}
\usage{
kwr_long_queries(kwr, longer_than = 60)
}
\arguments{
\item{kwr}{A kwresearch object.}

\item{longer_than}{A number.}
}
\value{
The output of \code{kwr_queries} filtered for queries longer than
\code{longer_than} characters.
}
\description{
Outputs queries longer than a specific number of characters
}
