% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngrams.R
\name{kwr_ngrams}
\alias{kwr_ngrams}
\title{Generate n-grams from queries}
\usage{
kwr_ngrams(
  queries,
  max_words = 4,
  min_words = 1,
  min_n = 1,
  min_volume = 1,
  order = volume
)
}
\arguments{
\item{queries}{A data frame containing queries and their search volumes.
Should be the result of functions kwr_source_queries, kwr_clean_queries,
kwr_classified_queries, or kwr_valid_queries.}

\item{max_words}{Maximum number of words in n-grams.}

\item{min_words}{Minimum number of words in n-grams.}

\item{min_n}{Minimum number of queries. Only the n-grams with at least this
number of queries will be included.}

\item{min_volume}{Minimum search volume per n-gram. Only the n-grams with at
least this volume will be included.}

\item{order}{A column the result will be reverse ordered by.}
}
\value{
A tibble of n-grams with a basic stats (nuber of queries and sum of
search volumes).
}
\description{
Generate n-grams from queries
}
\examples{
queries <- data.frame(
  query = c("seo", "keyword research"),
  volume = c(1000, 500)
)
kwr <- kwresearch(queries)
kwr |> kwr_clean_queries() |>
  kwr_ngrams()
}
