% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngrams.R
\name{kwr_ngrams}
\alias{kwr_ngrams}
\title{Generates n-grams from queries}
\usage{
kwr_ngrams(
  kwr,
  max_words = 4,
  min_words = 1,
  min_n = 1,
  min_volume = 0,
  remove_nested = TRUE
)
}
\arguments{
\item{kwr}{A kwresearch object, which queries will be n-grams calculated
from.}

\item{max_words}{Maximum number of words in n-grams.}

\item{min_words}{Minimum number of words in n-grams.}

\item{min_n}{Minimum number of queries. Only the n-grams with at least this
number of queries will be included.}

\item{min_volume}{Minimum search volume per n-gram. Only the n-grams with at
least this volume will be included.}

\item{remove_nested}{If TRUE, n-grams fully contained in another n-gram
(number of queries must be the same) are filtered out from the result. May
be slower.}
}
\value{
A tibble of n-grams with a basic stats (nuber of queries and sum of
search volumes). The n-grams are orderd descendingly by number of queries
and search volume. Use dplyr::arrange to change order. If stop words are
set with the kwr_use_stopwords, they are removed from unigrams.
}
\description{
Generates n-grams from queries
}
\examples{
queries <- data.frame(
  query = c("seo", "keyword research"),
  volume = c(1000, 500)
)
kwr <- kwresearch(queries)
kwr |> kwr_ngrams()
}
