% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngrams.R
\name{kwr_ngrams}
\alias{kwr_ngrams}
\title{Generate n-grams from queries}
\usage{
kwr_ngrams(kwr, max_words = 4, min_words = 1, min_n = 1, min_volume = 1)
}
\arguments{
\item{kwr}{A kwresearch object, which clean queries will be n-grams
calculated from.}

\item{max_words}{Maximum number of words in n-grams.}

\item{min_words}{Minimum number of words in n-grams.}

\item{min_n}{Minimum number of queries. Only the n-grams with at least this
number of queries will be included.}

\item{min_volume}{Minimum search volume per n-gram. Only the n-grams with at
least this volume will be included.}
}
\value{
A tibble of n-grams with a basic stats (nuber of queries and sum of
search volumes). The n-grams are orderd descendingly by number of queries
and search volume. Use dplyr::arrange to change order.
}
\description{
Generate n-grams from queries
}
\examples{
queries <- data.frame(
  query = c("seo", "keyword research"),
  volume = c(1000, 500)
)
kwr <- kwresearch(queries)
kwr |> kwr_ngrams()
}
