% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune.R
\name{kwr_prune}
\alias{kwr_prune}
\title{Removes unwanted queries according to YAML recipes}
\usage{
kwr_prune(kwr, recipe_file, quiet = FALSE)
}
\arguments{
\item{kwr}{A kwresearch object to be pruned.}

\item{recipe_file}{A path to a recipe file in YAML format.}

\item{quiet}{If TRUE prints no messages.}
}
\value{
A kwresearch object with pruned data and status = 'pruned'.
}
\description{
If the status of the kwresearch object is 'data', it changes to 'pruned' and
pruned data is created from clean data. I pruned data already exists (i.e.
status = 'pruned') then pruned data are pruned again.
}
\examples{
\dontrun{
kwr <- kwr |> kwr_prune("prune.yml")
}
}
