% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kwresearch.R
\name{kwr_pruned_queries}
\alias{kwr_pruned_queries}
\title{Outputs pruned queries}
\usage{
kwr_pruned_queries(kwr)
}
\arguments{
\item{kwr}{A kwresearch object.}
}
\value{
A tibble.
}
\description{
Outputs pruned queries
}
