% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kwresearch.R
\name{kwr_queries}
\alias{kwr_queries}
\title{Outputs the most processed queries}
\usage{
kwr_queries(kwr)
}
\arguments{
\item{kwr}{A kwresearch object.}
}
\value{
A tibble with queries.
}
\description{
Outputs the most processed queries
}
