% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kwresearch.R
\name{kwr_removed_queries}
\alias{kwr_removed_queries}
\title{Outputs queries removed by kwr_prune}
\usage{
kwr_removed_queries(kwr, q = NULL)
}
\arguments{
\item{kwr}{A kwresearch object.}

\item{q}{A regular expression to filter the output data frame.}
}
\value{
A tibble with queries removed by kwr_prune.
}
\description{
Outputs queries removed by kwr_prune
}
