% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngrams.R
\name{kwr_subqueries}
\alias{kwr_subqueries}
\title{Generates n-grams from queries and filter only those that match any existing
query}
\usage{
kwr_subqueries(x, max_words = 5, min_words = 1, min_n = 1, min_volume = 0)
}
\arguments{
\item{x}{A kwresearch object, which queries will be n-grams calculated from,
or a data frame of queries and volume.}

\item{max_words}{Maximum number of words in n-grams.}

\item{min_words}{Minimum number of words in n-grams.}

\item{min_n}{Minimum number of queries. Only the n-grams with at least this
number of queries will be included.}

\item{min_volume}{Minimum search volume per n-gram. Only the n-grams with at
least this volume will be included.}
}
\value{
A tibble of n-grams with a basic stats (number of queries and sum of
search volumes). The n-grams are ordered descendingly by number of queries
and search volume. Use dplyr::arrange to change order.
}
\description{
In other words, it lists queries that are contained in other
queries.
}
