% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kwresearch.R
\name{kwr_summary}
\alias{kwr_summary}
\title{Shows number of queries in all phases of processing}
\usage{
kwr_summary(kwr)
}
\arguments{
\item{kwr}{A kwresearch object.}
}
\value{
Non.
}
\description{
Shows number of queries in all phases of processing
}
