% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex.R
\name{kwr_test_regex}
\alias{kwr_test_regex}
\title{Explore a regex pattern applied on queries}
\usage{
kwr_test_regex(
  kwr,
  pattern,
  and = NULL,
  except = NULL,
  stopwords = kwr_stopwords()
)
}
\arguments{
\item{kwr}{A \code{\link{kwresearch}} object.}

\item{pattern}{A single regular expression.}

\item{and}{A vector of regular expressions. If more than one, they are
treated as OR.}

\item{except}{A vector of regular expressions. If more than one, they are
treated as OR.}

\item{stopwords}{A vector of stopwords to remove from n-grams.}
}
\value{
A list of data frames.
}
\description{
Explore a regex pattern applied on queries
}
