% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kwresearch.R
\name{kwresearch}
\alias{kwresearch}
\title{Creates an object of the kwresearch class}
\usage{
kwresearch(queries = NULL, accentize = TRUE, normalize = TRUE)
}
\arguments{
\item{queries}{Queries to import as a data frame with at least two colums:
query (string) and volume (numeric). Optionally the data frame can contain
three additional columns: cpc (double), input (char) and source (char). If
you don't provide this argument, you can import queries later with
kwr_import or kwr_import_mm.}

\item{accentize}{The import functions tries to add correct accents (diakritic
marks) to queries without them.}

\item{normalize}{The import funkctions tries to unite queries, which differ
only by order of words.}
}
\value{
An object of the kwresearch class.
}
\description{
An object of the kwresearch class is a home to your keyword
research. You can import datasets into it, research queries, clasify them
and export the final keyword analysis.
}
\examples{
# create an empty object

kwr <- kwresearch()

# create an object and import queries

queries <- data.frame(
  query = c("seo", "keyword research"),
  volume = c(1000, 500)
)
kwr <- kwresearch(queries)
}
