% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kwresearch.R
\name{kwresearch}
\alias{kwresearch}
\title{Creates an object of the kwresearch class}
\usage{
kwresearch(queries = NULL)
}
\arguments{
\item{queries}{Queries to import as a data frame with at leas two colums:
query (string) and volume (numeric). Optionally the data frame can contain
three additional columns: cpc (double), input (char) and source (char).}
}
\value{
An object of the kwresearch class.
}
\description{
An object of the kwresearch class is a home to your keyword
research. You can import datasets into it, research queries, clasify them
and export the final keyword analysis.
}
\examples{
kwr <- kwresearch()

queries <- data.frame(
  query = c("seo", "keyword research"),
  volume = c(1000, 500)
)
kwr <- kwresearch(queries)
}
