% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune.R
\name{process_prune_recipe}
\alias{process_prune_recipe}
\title{Processes a single pruning recipe}
\usage{
process_prune_recipe(df, recipe, quiet = FALSE)
}
\arguments{
\item{df}{A data frame from a kwr object (either clean_data, or
pruned_data).}

\item{recipe}{A single pruning recipe (type 'remove').}

\item{quiet}{If TRUE prints no messgaes.}
}
\value{
A data frame with pruned queries.
}
\description{
Processes a single pruning recipe
}
\keyword{internal}
